refactor(backend:session): 将游戏连接逻辑收归会话
